+++ v-next/hardhat-errors/src/descriptors.ts
@@ -863,6 +863,14 @@ Please double check your arguments.`,
 
 Please double check your script's path.`,
       },
+      INVALID_FILE_PATH: {
+        number: 601,
+        messageTemplate: `The path "{path}" is a directory. A file path was expected.`,
+        websiteTitle: "Invalid file path",
+        websiteDescription: `A file path was expected, but a directory was provided.
+
+Please specify a valid file path.`,
+      },
     },
     NETWORK: {
       INVALID_URL: {